feat(pre-commit-install): Install default hooks
Pre-commit 2.18.0 introduced default_install_hook_types, allowing
.pre-commit-config.yaml files to specify the hook types to install by
default rather than requiring them to be passed on the command line.
In light of this new feature, only install default hooks now. This is a
breaking change for repositories relying on the explicit list of hooks
we happened to install previously: commit-msg, post-checkout,
pre-commit, pre-merge-commit, and pre-push.