fix(pre-commit): Remove --loose from license audit
The --loose argument was previously required to inspect the contents of
license files in packages that didn't properly declare their license
in their package.json. We are no longer using any such packages, but the
args can be overridden as needed. Also, run the yarn-audit-licenses hook
when the license config file, .licenses.config.ts, is modified.