feat(MegaLinter): Add v6 hooks at runner v6.7.1
Add v6 megalinter-incremental hook that only runs on the modified files,
skipping all linters that run in project mode. This is an upgraded
version of the v5 megalinter hook. Add v6 megalinter-full hook that runs
only on the modified files when supported, but also runs linters that
run in project mode. This is an upgraded version of the v5
megalinter-all hook. This brings MegaLinter's pre-commit hooks more in
line with typical hooks, which run on the files passed by pre-commit
rather than running git diff themselves.

Leverage new v6 features to automatically clean up the Docker container
and megalinter-reports directory, and give a meaningful name to the
Docker container rather than one randomly generated by Docker.